+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Can save metadata with `QuantifyExperiment` (!355)
 - Contribution guidelines updated, added: Versioning, Backward Compatibility and Deprecation Policy (!282)
+- MeasurementControl - Fix `print_progress` for 0 progress (!363)
 
 ## 0.6.2 (2022-06-30)